// options-cockpit/src/models/MarketSnapshot.ts
export interface MarketSnapshot {
    niftySpot: number;
    niftyOpen: number;
    niftyPreviousClose: number;
    niftyDayHigh: number;
    niftyDayLow: number;

    sensexSpot: number;
    sensexOpen: number;
    sensexPreviousClose: number;
    sensexDayHigh: number;
    sensexDayLow: number;

    indiaVix: number;
}